uhd: allow for 16 bit i2c and eeprom addrs

+++ host/include/uhd/types/serial.hpp
@@ -55,7 +55,7 @@ namespace uhd{
          * \param buf the vector of bytes
          */
         virtual void write_i2c(
-            boost::uint8_t addr,
+            boost::uint16_t addr,
             const byte_vector_t &buf
         ) = 0;
 
@@ -66,7 +66,7 @@ namespace uhd{
          * \return a vector of bytes
          */
         virtual byte_vector_t read_i2c(
-            boost::uint8_t addr,
+            boost::uint16_t addr,
             size_t num_bytes
         ) = 0;
 
@@ -77,8 +77,8 @@ namespace uhd{
          * \param buf the vector of bytes
          */
         virtual void write_eeprom(
-            boost::uint8_t addr,
-            boost::uint8_t offset,
+            boost::uint16_t addr,
+            boost::uint16_t offset,
             const byte_vector_t &buf
         );
 
@@ -90,8 +90,8 @@ namespace uhd{
          * \return a vector of bytes
          */
         virtual byte_vector_t read_eeprom(
-            boost::uint8_t addr,
-            boost::uint8_t offset,
+            boost::uint16_t addr,
+            boost::uint16_t offset,
             size_t num_bytes
         );
     };
