+++ host/lib/usrp/cores/i2c_core_100_wb32.cpp
@@ -72,7 +72,7 @@ public:
     }
 
     void write_i2c(
-        boost::uint8_t addr,
+        boost::uint16_t addr,
         const byte_vector_t &bytes
     ){
         _iface->poke32(REG_I2C_DATA, (addr << 1) | 0); //addr and read bit (0)
@@ -95,7 +95,7 @@ public:
     }
 
     byte_vector_t read_i2c(
-        boost::uint8_t addr,
+        boost::uint16_t addr,
         size_t num_bytes
     ){
         byte_vector_t bytes;
@@ -121,7 +121,7 @@ public:
 
     //override read_eeprom so we can write once, read all N bytes
     //the default implementation calls read i2c once per byte
-    byte_vector_t read_eeprom(boost::uint8_t addr, boost::uint8_t offset, size_t num_bytes)
+    byte_vector_t read_eeprom(boost::uint16_t addr, boost::uint16_t offset, size_t num_bytes)
     {
         this->write_i2c(addr, byte_vector_t(1, offset));
         return this->read_i2c(addr, num_bytes);
